Move blob bg code to hook
